+++ src/app/features/create-event/create-event.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
-import { CreateEventComponent } from './components/create-event/create-event.component';
+import { CreateEventComponent } from './create-event/create-event.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatInputModule } from '@angular/material/input';
-import { EventTypeFormComponent } from './components/event-type-form/event-type-form.component';
-import { EventDetailsFormComponent } from './components/event-details-form/event-details-form.component';
+import { EventTypeFormComponent } from './event-type-form/event-type-form.component';
+import { EventDetailsFormComponent } from './event-details-form/event-details-form.component';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatNativeDateModule } from '@angular/material/core';
