+++ src/app/features/create-event/create-event/create-event.component.html
@@ -1,0 +1,1 @@
+<router-outlet></router-outlet>
