+++ src/app/features/create-event/event-details-form/event-details-form.component.scss
