+++ src/app/features/create-event/event-details-form/hourErrorStateMatcher.ts
@@ -1,0 +1,10 @@
+import { FormControl, FormGroupDirective, NgForm } from '@angular/forms';
+import { ErrorStateMatcher } from '@angular/material/core';
+
+export class HourErrorStateMatcher implements ErrorStateMatcher {
+  isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
+    const isSubmitted = form && form.submitted;
+    if (form?.hasError('isHourInThePast')) return true;
+    return !!(control && control.invalid && (control.dirty || control.touched || isSubmitted));
+  }
+}
