+++ src/app/features/create-event/event-type-form/event-type-form.component.scss
@@ -1,0 +1,45 @@
+@use '@angular/material' as mat;
+@use '../../../styles/palettes' as palette;
+
+.event-type-form {
+  width: 100%;
+
+  &__title {
+    text-align: center;
+    width: 256px;
+  }
+
+  &__toggle-container {
+    background-color: mat.get-color-from-palette(palette.$gray-palette, 700);
+    min-height: 156px;
+    width: 100%;
+    margin-top: 64px;
+    padding: 16px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+  }
+
+  &__question {
+    width: 100%;
+    max-width: 700px;
+    text-align: center;
+    margin-bottom: 16px;
+  }
+
+  &__button-container {
+    width: 100%;
+    max-width: 700px;
+    display: flex;
+    justify-content: space-between;
+    margin-top: 64px;
+    padding: 0 8px;
+  }
+
+  &__button {
+    &--hidden {
+      visibility: hidden;
+    }
+  }
+}
