+++ src/app/features/create-event/index.ts
@@ -1,0 +1,2 @@
+export * from './create-event.module';
+export * from './create-event.interface';
