--- src/app/features/event/components/create-event/create-event.component.html
@@ -1,3 +1,0 @@
-<div class="create-event">
-  <router-outlet></router-outlet>
-</div>
