--- src/app/features/event/components/create-event/create-event.component.scss
@@ -1,7 +1,0 @@
-@use '@angular/material' as mat;
-@import '../../../../styles/_palettes.scss';
-
-.create-event {
-  display: flex;
-  flex-direction: column;
-}
