--- src/app/features/event/components/event-type-form/event-type-form.component.scss
@@ -1,43 +1,0 @@
-@use '@angular/material' as mat;
-@use '../../../../styles/palettes' as palette;
-
-.event-type-form {
-  width: 100%;
-}
-
-.event-type-form__title {
-  text-align: center;
-  width: 256px;
-}
-
-.event-type-form__toggle-container {
-  background-color: mat.get-color-from-palette(palette.$gray-palette, 700);
-  min-height: 156px;
-  width: 100%;
-  margin-top: 64px;
-  padding: 16px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}
-
-.event-type-form__question {
-  width: 100%;
-  max-width: 700px;
-  text-align: center;
-  margin-bottom: 16px;
-}
-
-.event-type-form__button-container {
-  width: 100%;
-  max-width: 700px;
-  display: flex;
-  justify-content: space-between;
-  margin-top: 64px;
-  padding: 0 8px;
-}
-
-.event-type-form__button--hidden {
-  visibility: hidden;
-}
